lpr-service/requirements.txt
# LPR Service Requirements
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
prometheus-client==0.21.0
requests==2.32.3
numpy==2.0.1
Pillow==10.4.0
